+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changes, deprecations and removals
 
+* Disable Cruise Control network resource goals when resource capacities are not set.
 * **From Strimzi 0.48.0 on, we support only Kubernetes 1.27 and newer.**
   Kubernetes 1.25 and 1.26 are not supported anymore.
 * Make properties `broker.session.timeout.ms`, `broker.heartbeat.interval.ms` and `controller.socket.timeout.ms` configurable